requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
psycopg2-binary
pyserial